requirements.txt
torch 
numpy 
tqdm
matplotlib
jupyter
pyyaml
mlflow